Port to Minecraft 1.21

+++ src/main/java/me/jellysquid/mods/sodium/mixin/features/render/immediate/matrix_stack/VertexConsumerMixin.java
@@ -17,7 +17,7 @@ public interface VertexConsumerMixin {
     VertexConsumer normal(float x, float y, float z);
 
     @Shadow
-    VertexConsumer vertex(double x, double y, double z);
+    VertexConsumer vertex(float x, float y, float z);
 
     /**
      * @reason Avoid allocations
